pyproject.toml
[project]
name = "map-service"
version = "0.1.0"
description = ""
authors = [
    {name = "JulianAVG64",email = "julianavg2002@gmail.com"}
]
readme = "README.md"
requires-python = "^3.13"
dependencies = [
    "fastapi[all] (>=0.115.12,<0.116.0)",
    "uvicorn (>=0.34.3,<0.35.0)",
    "requests (>=2.32.4,<3.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
black = "^25.1.0"
isort = "^6.0.1"

